+++ CHANGELOG.md
@@ -6,8 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 [Unreleased]
 ### Added
+- Warning is emitted if bootstrap expansion factor is too large
+- Warning is emitted if maxcall/maxiter stop the iterations too early
 ### Changed
+- The clustering/K-means for ellipsoid decomposition is now done in scaled space of points divided by stddev along each dimension
 ### Fixed
+- Fix a rare possible bug where due to numerical instability in uniform sampling a q==0 error could have occured
 
 ## [2.1.2] - 2023-06-01